// src/app/api/model/eFormularDTO.ts
/**
 * Next-Gen App API
 * Die REST Web-Schnittstelle zu der Next-Gen App Anwendung
 *
 * The version of the OpenAPI document: v1
 * Contact: info@brunnerinformatik.ch
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EFormularAntwortKanal } from './eFormularAntwortKanal';
import { EFormularTypDTO } from './eFormularTypDTO';
import { EDokumentDTO } from './eDokumentDTO';
import { EFormularDokumentPoolDTO } from './eFormularDokumentPoolDTO';
import { EFormularStatus } from './eFormularStatus';


/**
 * Diese Klasse reprÃ¤sentiert ein Data-Transfer-Object fÃ¼r Elektro-Formular EntitÃ¤ten.
 */
export interface EFormularDTO { 
    /**
     * Die <see cref=\"P:BrunnerInformatik.NextGen.App.ViewModel.EFormularDTO.Guid\">GUID</see> des Mandanten, in welchem sich dieses Formular befindet.
     */
    mandant: string;
    /**
     * Die <see cref=\"P:BrunnerInformatik.NextGen.App.ViewModel.EFormularDTO.Guid\">GUID</see> dieses Formulars.
     */
    guid?: string | null;
    dokument?: EDokumentDTO;
    formularTyp?: EFormularTypDTO;
    /**
     * Die <see cref=\"P:BrunnerInformatik.NextGen.App.ViewModel.EFormularDTO.Guid\">GUID</see> des EmpfÃ¤ngers dieses Formulars.
     */
    empfaenger?: string | null;
    fortschritt?: number;
    status?: EFormularStatus;
    antwortKanal?: EFormularAntwortKanal;
    antwortEMail?: string | null;
    antwortM20?: string | null;
    formularDokumentPool?: Array<EFormularDokumentPoolDTO> | null;
}