// src/app/api/model/eFormularBeilageDTO.ts
/**
 * Next-Gen App API
 * Die REST Web-Schnittstelle zu der Next-Gen App Anwendung
 *
 * The version of the OpenAPI document: v1
 * Contact: info@brunnerinformatik.ch
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EFormularTypDTO } from './eFormularTypDTO';
import { EDokumentDTO } from './eDokumentDTO';


/**
 * Diese Klasse reprÃ¤sentiert ein Formularbeilage fÃ¼r ein Formular aus der Elektro-Sparte.
 */
export interface EFormularBeilageDTO { 
    /**
     * Die <see cref=\"P:BrunnerInformatik.NextGen.App.ViewModel.EFormularBeilageDTO.Guid\">GUID</see> des Mandanten, in welchem sich die Formularbeilage befindet.
     */
    mandant: string;
    /**
     * Die <see cref=\"P:BrunnerInformatik.NextGen.App.ViewModel.EFormularBeilageDTO.Guid\">GUID</see> dieser Formularbeilage.
     */
    guid?: string | null;
    /**
     * Die <see cref=\"P:BrunnerInformatik.NextGen.App.ViewModel.EFormularBeilageDTO.Guid\">GUID</see> des Formulars, zu welchem diese Formularbeilage gehÃ¶rt.
     */
    formular?: string | null;
    dokument?: EDokumentDTO;
    formularTyp: EFormularTypDTO;
}